Create recurse_crawl_download.js
If you want to modify the script to download the content of each page, you would need to add a function to write the response body to a file. Here's how you could modify the callback function to save the HTML content of each page:

With this modification, the script will save the HTML content of each page to a file in a directory named downloaded within the directory where the script is located. The filename is generated by replacing all non-alphanumeric characters in the URL with underscores and appending .html to the end.

Remember to handle the crawling and saving of content responsibly and respect the robots.txt rules and terms of service of the websites you are crawling.